\name{annoTable}
\alias{annoTable}
\title{Annotation results}
\description{Shows the table of annotation results by eveRest annotation}
\usage{
annoTable(Experiment)
}
\arguments{
  \item{Experiment}{A 'MetaboSet' S4 object containing the experiment data previously created by evAnnotate.}
  }
\details{
See eveRest vignette for more details. To open the vignette, execute the following code in R:
vignette("eveRestManual", package="everest")

}
\examples{
	#annoTable(ex) 
}

