\name{hmdb}
\alias{hmdb}
\title{Human Metabolome Database}
\description{The default metabolomics library of eveRest.}
\usage{
data(hmdb)
}
\details{
This is the eveRest default library, and automatically loaded with the eveRest package. For details, see reference below.
}
\references{
[1] Wishart, D. S.; Jewison, T.; Guo, A. C.; Wilson, M.; Knox, C.; Liu, Y.; Djoumbou, Y.; Mandal, R.; Aziat, F.; Dong, E.; Bouatra, S.; Sinelnikov, I.; Arndt, D.; Xia, J.; Liu, P.; Yallou, F.; Bjorndahl, T.; Perez-Pineiro, R.; Eisner, R.; Allen, F.; Neveu, V.; Greiner, R.; Scalbert, A. HMDB 3.0-The Human Metabolome Database in 2013. Nucleic Acids Res. (2013), 41, 801-807.
 
}
